Added dom_element_to_compare.
Using this parameter, it's possible to compare the text content of a DOM
element when computing similarity with content on disk.  This way, the
comparison can be more precise than when comparing the whole document.
This parameter can be used for particular domain.  If the DOM element is
not found in the content, the default behavior of comparing the whole
document will be used instead.