Stop processing too many databases
Recently, several databases of old tweets are being made available.
This represents a lot of data to process which takes so much time that that the latest database is not processed on time or frequently enough.

To prevent this issue, a maximum of 3 databases are processed per
iteration: the latest database + 2 older databases.

Eventually, all the old databases will be processed while processing the
latest database frequently.